+++ infrastructure/src/main/java/mops/controllers/DozentenRepository.java
@@ -1,7 +1,7 @@
 package mops.controllers;
 
 import java.util.List;
-import mops.Dozent;
+import mops.rollen.Dozent;
 
 public interface DozentenRepository {
   public List<Dozent> getAll();
