+++ infrastructure/src/main/java/mops/controllers/FragebogenRepository.java
@@ -3,7 +3,7 @@ package mops.controllers;
 import java.time.LocalDateTime;
 import java.util.List;
 import mops.Fragebogen;
-import mops.SkalarFrage;
+import mops.fragen.SkalarFrage;
 import mops.fragen.TextFrage;
