+++ infrastructure/src/main/java/mops/database/MockFragebogenRepository.java
@@ -7,12 +7,12 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
-import mops.Auswahl;
+import mops.fragen.Auswahl;
 import mops.Einheit;
 import mops.Frage;
 import mops.Fragebogen;
 import mops.fragen.MultipleChoiceFrage;
-import mops.SkalarFrage;
+import mops.fragen.SkalarFrage;
 import mops.fragen.TextFrage;
 import mops.controllers.FragebogenRepository;
 import org.springframework.beans.factory.annotation.Qualifier;
