+++ model/src/main/java/mops/Fragebogen.java
@@ -21,7 +21,7 @@ public class Fragebogen {
   private final Long bogennr;
   private String veranstaltungsname;
   private String professorenname;
-  private List<Frage> fragen;
+  private List<mops.Frage> fragen;
   private LocalDateTime startdatum;
   private LocalDateTime enddatum;
   private Einheit type;
