+++ model/src/main/java/mops/Kontaktformular.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import mops.rollen.Student;
 
 @Getter
 @Setter
