+++ model/src/main/java/mops/antworten/TextAntwort.java
@@ -2,7 +2,6 @@ package mops.antworten;
 
 import lombok.AllArgsConstructor;
 import lombok.Value;
-import mops.antworten.Antwort;
 
 @Value
 @AllArgsConstructor
