+++ model/src/main/java/mops/fragen/MultipleChoiceFrage.java
@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.Map;
 import lombok.Getter;
 import lombok.Setter;
-import mops.Auswahl;
 import mops.Frage;
 import mops.antworten.MultipleChoiceAntwort;
