+++ model/src/main/java/mops/fragen/MultipleResponseFrage.java
@@ -1,6 +1,5 @@
 package mops.fragen;
 
-import mops.fragen.MultipleChoiceFrage;
 
 public class MultipleResponseFrage extends MultipleChoiceFrage {
