+++ model/src/main/java/mops/rollen/Dozent.java
@@ -1,13 +1,12 @@
-package mops;
+package mops.rollen;
 
 import lombok.Value;
 
 @Value
-public class Student {
+public class Dozent {
 
   private Long id;
   private String vorname;
   private String nachname;
   private String email;
-  private Integer matnr;
 }
