+++ model/src/main/java/mops/rollen/Student.java
@@ -1,12 +1,13 @@
-package mops;
+package mops.rollen;
 
 import lombok.Value;
 
 @Value
-public class Dozent {
+public class Student {
 
   private Long id;
   private String vorname;
   private String nachname;
   private String email;
+  private Integer matnr;
 }
