+++ services/src/main/java/mops/TypeChecker.java
@@ -2,6 +2,7 @@ package mops;
 
 
 import mops.fragen.MultipleChoiceFrage;
+import mops.fragen.SkalarFrage;
 import mops.fragen.TextFrage;
 
 public class TypeChecker {
